the road to 1.0, first stretch: keys, interface, discovery, limits, power loss, the last release
What 1.0 must guarantee is now a list in the README with a test behind
each line, and the pieces that list needed are in. Version 0.9.7.

Release keys (SECURITY.md, MANUAL 13.8): two are built into the kernel,
one that signs releases and one kept apart against the loss of the
first; a package signed with either is accepted. A signed rotation --
tools/sign-rotation.sh, published as the asset "rotate", EBROTATE in
formats.h -- moves a machine to one key, written into its settings as
"release key |"; from then on that key alone is trusted and a package
signed with the previous key is refused. Read on every check before the
version is compared. tools/rotate-test.sh proves the three cases.
tools/sign-release.sh also writes PROVENANCE (commit, day, toolchain).

The program interface (MANUAL 18): entry, the eight calls and their
registers, the message layout, the tags, the image format. sdk/erebus.h
and sdk/hello.c are a program from outside the tree; tools/sdktest.sh
sends them through the door, compiles hello.c beside its header on the
machine, runs it and gives it a text.

Discovery (MANUAL 13.3): "discovery | local" is the default and answers
a SEEK from the own network and from known nodes; "open", "known" and
"quiet" widen or narrow it. A machine not in the nodes table is told
the name and the work flag only; the version, free memory, uptime, job
count, key and the addresses heard are for known nodes. Gossip is taken
only from a machine the setting would answer. A SEEK left unanswered
is said on the console once a minute per address. tools/pipe-local.sh
probes with a machine that is not EreBUS (tools/pipe-probe.py, raw
frames on a multicast wire); pipe-two runs under the setting.

That probe found a bug: ip_send took the address to answer as a pointer
into frame_in, and mac_for pumps the wire while it waits for the
neighbour's ARP answer, which lands in frame_in and turned the address
into bytes of the answer (153.153.10.9). The first reply to any new
neighbour went there; only the cache filled on that failed try let the
second through. ip_send copies the address first now.

With that first reply arriving, a second thing showed: knock_claims
took the name and version from a signed handshake only when the heard
entry had none, so a machine that announced itself before its name was
written stayed under the default name in the row its handshake then
made, and "allow <name>" found no such node (pipe-vouch). The signed
claim overwrites the unsigned one now.

Limits: MAX_CPUS is 64, in eb/smp.h alone; the MADT's type-9 (x2apic)
entries are read beside the type-0 ones, an id above 255 needs x2apic
mode and is named when the firmware left it off, processors beyond the
build are counted and named; the processor masks are 64 bits. The
per-processor fault stacks are no longer static for every processor the
kernel is built for -- sixty-four sets were three megabytes of image
and did not fit the loader's room at 2 MiB -- an application processor
takes its three from the frame allocator as it comes up. NODES_MAX is
64 and the table's text 16 KiB; a 4 KiB table from an older store is
replaced by a wider one on the way in, its rows carried over. HEARD_MAX
is 64. tools/limits.sh boots -smp 64 and trusts 64 nodes through the
door, the 65th refused.

Power loss: tools/powerloss.sh kills QEMU outright a moment into a
write of the store, the writes throttled so the moment is wide; every
other boot lets a write finish so the ring goes round. Every boot must
restore a generation no older than the last seen finished and never the
one the cut fell into (a generation is claimed by its header, written
last); the cuts landing inside a write are counted, and a run with none
proves nothing and fails. A dozen in the battery, a hundred in the
release gate. main.c says "snap: writing generation N" before a write,
so a log cut short shows one in flight.

The last release: tools/oldbuild.sh builds a released tag once from a
git worktree; tools/oldstore.sh boots it on a fresh store and then
today's kernel on that store; tools/oldpipe.sh runs a node of each and
sends an object across. Both in CI and in the battery's serial part.

Gates: tools/selfkernel.sh exits non-zero on failure and the gate uses
it; tools/kvm-battery.sh adds a hundred power cuts; .github/workflows/
nightly.yml runs the gate at 03:00 UTC on a self-hosted runner with KVM
(MANUAL 15 says how to register one); CI builds with -Werror and runs
the compatibility and interface tests under TCG; badges in the README;
.clang-format describes the style without reformatting the tree.

Docs: README "What 1.0 guarantees", the release entry; MANUAL 6, 13.3,
13.8, 13.9, 14, 15, 16, 17, 18; SECURITY.md on the keys. The nodes
table's widening and the settings' discovery change are said in the
journal.

Verified: battery 44 of 44 green under KVM (38 in lanes, 6 alone), the
seven new tests among them; powerloss with 24 runs round the ring; the
frame and trampoline guards pass. TASK_MAX, named as a limit in the
plan, is the combine selector "max", not a limit.

+++ kernel/include/eb/syscall.h
@@ -41,7 +41,7 @@
 void syscall_init(void);
 
 /* A bit per processor that has run a ring-3 system call. */
-u32 syscall_cpu_mask(void);
+u64 syscall_cpu_mask(void);
 
 /* percpu_init / percpu_set_kernel_stack live in eb/percpu.h, included above. */
